Move escapeMetricName to mapper
